+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [clients:middlewares] Use `toLowerObjectKeys` instead of ramad when changing headers object.
 
 ## [6.44.0] - 2021-07-15